requirements.txt
keyboard
py-hangul-utils
